LayoutTests/http/tests/security/XFrameOptions/x-frame-options-deny-multiple-clients.html

<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
Test that two main resources pointing to the same url that are canceled within didReceiveResponse() don't cause us to crash.

<iframe src="http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny.cgi" onload="alert('PASS: onload fired.');"></iframe>
<iframe src="http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny.cgi" onload="alert('PASS: onload fired.');"></iframe>
</body>
</html>
